Sum of round numbers

+++ Day_17/Learn_From_Math.java
@@ -2,7 +2,7 @@ import java.util.Scanner;
 public class Learn_From_Math {
     public static void main(String[] args){
         Scanner s = new Scanner(System.in);
-        int limit = 1e6+1;
+        int limit = (int)1e6+1;
         int[] ar = new int[limit];
 
     }
